+++ examples/rbfs/CMakeLists.txt
@@ -11,6 +11,6 @@
 ##
 ## ---------------------------------------------------------------------
 
-FOREACH(_dir ex0 ex1)
+FOREACH(_dir ex0)
   ADD_SUBDIRECTORY(${_dir})
 ENDFOREACH()
