--- examples/rbfs/ex1/README
@@ -1,2 +1,0 @@
-A test program to check if we can solve a system with augmented degrees of
-freedom.
