--- examples/rbfs/ex1/input2d
@@ -1,101 +1,0 @@
-RADIUS = 1
-N = 64
-L = 4.0
-DX = L  / N
-MFAC = 1.0
-ELEM_TYPE = "TRI3"
-XCOM = 0.0
-YCOM = 0.0
-D = 1.0
-DT = 0.5 * 4.0 / 256
-C = 1.0 / DT
-
-LaplaceOperator {
- dist_to_bdry = 4.0 * DX
- eps = 0.25 * DX
- C = C
- D = D
-}
-
-solver {
-
-}
-
-Q {
-   function = "sin(PI*X_0)*sin(PI*X_1)"
-}
-
-B {
-   c = C
-   d = D
-   function = "(c + 2.0 * d * PI * PI) * sin(PI * X_0) * sin(PI * X_1)"
-}
-
-ls {
-   r = RADIUS
-   xcom = XCOM
-   ycom = YCOM
-   function = "r - sqrt((X_0-xcom)*(X_0-xcom) + (X_1-ycom)*(X_1-ycom))"
-}
-
-Main {
-// log file parameters
-   log_file_name = "RBF.log"
-   log_all_nodes = FALSE
-
-// visualization dump parameters
-   viz_writer = "VisIt"
-   viz_dump_dirname = "viz2d"
-   visit_number_procs_per_file = 1
-
-// timer dump parameters
-   timer_enabled = TRUE
-}
-
-CartesianGeometry {
-   domain_boxes       = [(0,0), (N - 1,N - 1)]
-   x_lo               = -L/2, -L/2      // lower end of computational domain.
-   x_up               = L/2, L/2      // upper end of computational domain.
-   periodic_dimension = 1, 1
-}
-
-GriddingAlgorithm {
-   max_levels = 1                 // Maximum number of levels in hierarchy.
-
-   ratio_to_coarser {
-      level_1 = 4, 4              // vector ratio to next coarser level
-   }
-
-   largest_patch_size {
-      level_0 = 512, 512          // largest patch allowed in hierarchy
-                                  // all finer levels will use same values as level_0...
-   }
-
-   smallest_patch_size {
-      level_0 =   4,   4          // smallest patch allowed in hierarchy
-                                  // all finer levels will use same values as level_0...
-   }
-
-   efficiency_tolerance = 0.70e0  // min % of tag cells in new patch level
-   combine_efficiency   = 0.85e0  // chop box if sum of volumes of smaller
-                                  // boxes < efficiency * vol of large box
-}
-
-StandardTagAndInitialize {
-   tagging_method = "REFINE_BOXES"
-   RefineBoxes {
-      level_0 = [( N/4 , N/4 ),( N/2 - 1 , N/2 - 1 )] , [( N/2 , N/4 ),( 3*N/4 - 1 , N/2 - 1 )] , [( N/4 , N/2 ),( N/2 - 1 , 3*N/4 - 1 )]
-   }
-}
-
-LoadBalancer {
-   bin_pack_method = "SPATIAL"
-   max_workload_factor = 1
-}
-
-TimerManager{
-   print_exclusive = FALSE
-   print_total = TRUE
-   print_threshold = 1.0
-   timer_list = "IBTK::*::*"
-}
